ai-vision/requirements.txt
Flask==3.0.0
azure-identity==1.7.0
azure-storage-blob==12.8.1
python-dotenv==1.0.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.5.0
opencv-python==4.8.0.74
msrest==0.7.0
requests==2.31.0
gunicorn==21.2.0 